requirements.txt
requests>=2.31.0
tabulate>=0.9.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0